+++ pkg/analyzer/lib/src/dart/resolver/exit_detector.dart
@@ -171,6 +171,27 @@ class ExitDetector extends GeneralizingAstVisitor<bool> {
   }
 
   @override
+  bool visitDotShorthandConstructorInvocation(
+    DotShorthandConstructorInvocation node,
+  ) {
+    return _nodeExits(node.argumentList);
+  }
+
+  @override
+  bool visitDotShorthandInvocation(DotShorthandInvocation node) {
+    var element = node.memberName.element;
+    if (_elementExits(element)) {
+      return true;
+    }
+    return _nodeExits(node.argumentList);
+  }
+
+  @override
+  bool visitDotShorthandPropertyAccess(DotShorthandPropertyAccess node) {
+    return _elementExits(node.propertyName.element);
+  }
+
+  @override
   bool visitEmptyStatement(EmptyStatement node) => false;
 
   @override